custom wscripts are NOT supported by cloudpebble
at least not as of 2016-07-01.

Generate and store the files needed. Typically storing
generated files under source code control is not a good idea
but there is no alternative for CloudPebble at this time.

NOTE the generated js (long) file path is the same as what the
wscript creates, retaining as-is for compatibility for now.